+++ CHANGELOG.md
@@ -3,19 +3,34 @@
 ###### Breaking
 
 * [PBXGroup] `#new_static_library` has been replaced by the more versatile
-  `#new_product_ref_for_target`.
+  `#new_product_ref_for_target`.  
+  [@irrationalfab](https://github.com/irrationalfab)
 
 * [XCScheme] Overhauled interface to support multiple targets.  
   [@jasonprado](https://github.com/jasonprado), [#95](https://github.com/CocoaPods/Xcodeproj/pull/95)
 
 ###### Enhancements
 
-* Improvements to the generation of new targets.
+* [Project] The project can now recreate it schemes from scratch and optionally
+  hide them.  
+  [@irrationalfab](https://github.com/irrationalfab)
 
-* [PBXGroup] Now defaults to sorting by name.
+* Added support for booleans in the C extension which handles Property list
+  files.  
+  [@irrationalfab](https://github.com/irrationalfab)
+
+* Improvements to the generation of new targets.  
+  [@irrationalfab](https://github.com/irrationalfab)
+
+* [PBXGroup] Now defaults to sorting by name.  
+  [@irrationalfab](https://github.com/irrationalfab)
 
 * [XCScheme] The string representation of schemes now closely matches Xcode
-  behaviour.
+  behaviour.  
+  [@irrationalfab](https://github.com/irrationalfab)
+
+* [PBXGroup, PBXFileReference] Added `#parents`  
+  [@irrationalfab](https://github.com/irrationalfab)
 
 ## 0.11.1